+++ test/unit/rx-schema.test.ts
@@ -843,7 +843,14 @@ config.parallel('rx-schema.test.js', () => {
                 }
             });
 
-            const found = await col.find().where('fileInfo.watch.time').gt(-9999999999999999999999999999).sort('fileInfo.watch.time').exec();
+            console.log('aaaaaa');
+            const query = col.find()
+                .where('fileInfo.watch.time')
+                .gt(-9999999999999999999999999999)
+                .sort('fileInfo.watch.time');
+            console.dir(query.toJSON());
+            console.log('aaaaaa DONE');
+            const found = await query.exec();
             assert.strictEqual(found.length, 1);
             assert.strictEqual(found[0].fileInfo.watch.time, 1);